seperated query from display functions
